# packages/contracts/openapi.yaml
openapi: 3.1.0

info:
  title: Rehearsal control-plane
  version: 0.2.0
  description: |
    The operations the browser calls — `ARCHITECTURE.md` §6.2. Their count is
    asserted in `src/contracts.test.ts`, so this prose cannot drift from `paths`.

    Written before the API, per `CODE_INSTRUCTION.md`'s contract-first rule, and
    the source of truth for the orval-generated client. Nothing else in the
    product is HTTP: once a session starts, the browser talks to the worker over
    the LiveKit data channel instead (§6.3).

    Component schemas are `$ref`s into `schema/`, which is generated from the zod
    validators the server and the worker actually run. The spec cannot drift from
    the code because it does not restate it — `pnpm check:schema` fails on drift.

servers:
  - url: http://localhost:3000
    description: Local development.

security:
  - SessionCookie: []

tags:
  - name: me
  - name: courses
  - name: sessions
  - name: progress

paths:
  /api/me:
    get:
      tags: [me]
      summary: The signed-in learner's profile
      description: |
        In development only, a request without a session cookie bootstraps the
        fixed dev user and sets an HttpOnly, SameSite=Lax rehearsal_session JWT
        cookie. Production requires a valid cookie; bootstrap is disabled.
        Invalid presented credentials return 401. New profiles have onboarded=false.
      security:
        - SessionCookie: []
        - {}
      operationId: getMe
      responses:
        "200":
          description: The profile, including recalled patterns from past sessions.
          headers:
            Set-Cookie:
              description: Development bootstrap only; Secure outside local HTTP.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      tags: [me]
      summary: Update level and the two preferences
      description: |
        Saving level atomically sets onboarded=true, including Skip submitting a
        level. Preference-only updates do not complete onboarding. onboarded is
        server-controlled and cannot be patched. An empty patch is a no-op.
      operationId: patchMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilePatch"
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/courses:
    get:
      tags: [courses]
      summary: Every course, badged for this learner
      description: |
        `levels` orders and badges, it never hides (§4.4) — so this always
        returns the full set and `fit` says which one is on level.
      operationId: listCourses
      responses:
        "200":
          description: All courses, each with a `fit` badge for the caller's level.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseCard"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/courses/{id}/unlearn:
    post:
      tags: [courses]
      summary: Mark a learned course as unlearned
      operationId: unlearnCourse
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "204":
          description: Course learning state cleared.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/courses/{id}:
    get:
      tags: [courses]
      summary: One course, fully parsed
      operationId: getCourse
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: The parsed course — the same object `parseCourse` returns.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/sessions:
    post:
      tags: [sessions]
      summary: Start a session
      description: |
        Creates the room, mints a token scoped to it, and returns the patterns
        worth recalling. The browser never holds a LiveKit key, only this token.
      operationId: startSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSessionRequest"
      responses:
        "201":
          description: |
            The session id, the LiveKit grant, and `recalled` — the slips from
            previous sessions the coach should watch for (§5.5).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStart"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/sessions/{id}/end:
    post:
      tags: [sessions]
      summary: End the session and get the debrief
      description: |
        Reachable from every state, including before the learner has said
        anything (§5.7). Idempotent — a second call returns the same debrief
        rather than a second one.
      operationId: endSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: The debrief. Built once, then served from storage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debrief"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/sessions/{id}/debrief:
    get:
      tags: [sessions]
      summary: Re-read a finished session's debrief
      description: |
        The same object `endSession` returned, for the history screen and for a
        page reload mid-debrief.
      operationId: getDebrief
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: The stored debrief.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debrief"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"
        "409":
          description: The session has not ended yet, so there is no debrief to read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/progress:
    get:
      tags: [progress]
      summary: Every goal, with its attempts and state
      description: |
        The whole catalog, not only the courses with history: a goal the learner
        has not started is `not_started` rather than missing, which is what lets
        the screen show "what needs doing" without a second call.

        `met` is permanent — a later worse attempt does not take it away, and no
        field carries a trend. Goals are ordered `unfinished`, `not_started`,
        `met`, so the client never sorts.

        Assembled by `progressView` (`ARCHITECTURE.md` §5.8) from the sessions
        table and the metrics each session reported when it ended. A session that
        ended before it could report falls back to its wall-clock span and zeroed
        counters rather than being hidden.
      operationId: getProgress
      responses:
        "200":
          description: Goals in screen order, plus the count tiles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: rehearsal_session
      description: |
        Signed JWT resolved to the caller's userId. Cookie is HttpOnly and
        SameSite=Lax, Secure outside local HTTP. Browser mutations must pass
        the server's configured Origin policy. No provider keys reach the web.
  parameters:
    CourseId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z][a-z0-9-]*$"
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1

  schemas:
    # Named `$ref`s into `schema/`, which is generated from zod. Naming them here
    # is what gives orval a clean model name (`Debrief`, not `endSession200`),
    # and it keeps a single place to see every shape that crosses HTTP.
    Profile:
      $ref: "./schema/http/Profile.schema.json"
    ProfilePatch:
      $ref: "./schema/http/ProfilePatch.schema.json"
    Course:
      $ref: "./schema/course.schema.json"
    CourseCard:
      $ref: "./schema/http/CourseCard.schema.json"
    SessionStart:
      $ref: "./schema/http/SessionStart.schema.json"
    Debrief:
      $ref: "./schema/http/Debrief.schema.json"
    Progress:
      $ref: "./schema/http/Progress.schema.json"
    StartSessionRequest:
      $ref: "./schema/http/StartSessionRequest.schema.json"
    Error:
      $ref: "./schema/http/ApiErrorBody.schema.json"

  responses:
    Unauthorized:
      description: No valid session token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such course or session owned by the caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid parameters or JSON body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: A required dependency is unavailable; retry rather than inventing session data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error with a sanitized learner-readable message.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"